# docs/swagger.yaml
basePath: /
definitions:
...
  requests.DebugEchoRequest:
    properties:
      message:
        type: string
    type: object
  requests.ExecuteDiagnosisRequest:
    properties:
      conversation_data:
        additionalProperties: true
        type: object
      target_avatar_id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FBV
        type: string
    required:
    - conversation_data
    - target_avatar_id
    type: object
...
  response.AvatarAI:
    properties:
      bio:
        example: あなたの性格や好みを反映した分身AIです
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FDV
        type: string
      name:
        example: 私の分身AI
        type: string
      personality:
        additionalProperties:
          type: string
        example:
          ' "style"': ' "casual"}'
          '{"tone"': ' "friendly"'
        type: object
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      user_id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FAV
        type: string
    type: object
...
  response.DebugEchoResponse:
    properties:
      message:
        type: string
    type: object
  response.DebugIDTokenResponse:
    properties:
      expires_in:
        example: "3600"
        type: string
      id_token:
        example: eyJhbGciOiJSUzI1NiIsImtpZCI6IjFlOWdkazcifQ...
        type: string
      token_type:
        example: Bearer
        type: string
      user_id:
        example: user-12345
        type: string
    type: object
  response.DiagnosisDetail:
    properties:
      analysis_result:
        additionalProperties: true
        description: AI分析詳細
        type: object
      can_direct_chat:
        type: boolean
      conversation_data:
        additionalProperties: true
        description: 会話データ
        type: object
      created_at:
        type: string
      diagnosis_score:
        type: integer
      id:
        type: string
      points_earned:
        type: integer
      target_avatar:
        allOf:
        - $ref: '#/definitions/response.AvatarAI'
        description: 相手のAvatar詳細
      user_avatar:
        allOf:
        - $ref: '#/definitions/response.AvatarAI'
        description: 自分のAvatar詳細
    type: object
  response.DiagnosisDetailResponse:
    properties:
      diagnosis:
        $ref: '#/definitions/response.DiagnosisDetail'
    type: object
  response.DiagnosisHistory:
    properties:
      can_direct_chat:
        type: boolean
      created_at:
        type: string
      diagnosis_score:
        type: integer
      id:
        type: string
      points_earned:
        type: integer
      target_avatar_name:
        description: 相手のAvatar名
        type: string
      user_avatar_name:
        description: 自分のAvatar名
        type: string
    type: object
  response.DiagnosisHistoryResponse:
    properties:
      histories:
        items:
          $ref: '#/definitions/response.DiagnosisHistory'
        type: array
    type: object
  response.DiagnosisResult:
    properties:
      analysis_result:
        additionalProperties: true
        description: AI分析結果
        type: object
      can_direct_chat:
        description: 直接チャット可能フラグ
        type: boolean
      created_at:
        type: string
      diagnosis_id:
        type: string
      diagnosis_score:
        description: 1-5
        type: integer
      points_earned:
        description: 獲得ポイント
        type: integer
    type: object
  response.ErrorResponse:
    properties:
      error:
        example: invalid_token
        type: string
      message:
        example: Firebase IDトークンが不正または無効です
        type: string
    type: object
...
  response.GetUnlockStatusResponse:
    properties:
      status:
        $ref: '#/definitions/response.UnlockStatus'
    type: object
...
  response.Match:
    properties:
      chat_id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FAV
        type: string
      final_score:
        example: 85
        type: integer
      id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FEV
        type: string
      matched_at:
        example: "2024-01-01T12:00:00Z"
        type: string
      partner_id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FBV
        type: string
      partner_image_url:
        example: https://example.com/images/profile2.jpg
        type: string
      partner_name:
        example: 佐藤花子
        type: string
      user_id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FAV
        type: string
    type: object
...
  response.MatchResponse:
    properties:
      match:
        $ref: '#/definitions/response.Match'
      message:
        example: マッチングが成立しました！
        type: string
    type: object
...
  response.UnlockStatus:
    properties:
      can_match:
        description: マッチング可能かどうか
        example: false
        type: boolean
      chat_id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FAV
        type: string
      current_score:
        example: 75
        type: integer
      is_unlocked:
        example: false
        type: boolean
      remaining_score:
        description: アンロックまでに必要なスコア
        example: 5
        type: integer
      unlock_threshold:
        example: 80
        type: integer
    type: object
...
info:
  contact: {}
  license:
    name: MIT
  title: Hackathon API
  version: "1.0"
paths:
...
  /chats/{partnerUserId}/match:
    post:
      description: マッチング条件を達成した場合にマッチングを成立させる（サーバー側で自動判定、Firestore通知）
      parameters:
      - description: 相手のユーザーID（ULID）
        in: path
        name: partnerUserId
        required: true
        type: string
      responses:
        "200":
          description: マッチング成立成功
          schema:
            $ref: '#/definitions/response.MatchResponse'
        "400":
          description: マッチング条件を満たしていない
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: このチャットにアクセスする権限がない、または既にマッチング済み
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: チャットが見つからない
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: マッチング成立（サーバー側で自動判定）
      tags:
      - matches
...
  /chats/{partnerUserId}/unlock-status:
    get:
      description: 特定チャットのアンロック状態（マッチング可能かどうか）を取得する（IDは相手のUserID）
      parameters:
      - description: 相手のユーザーID（ULID）
        in: path
        name: partnerUserId
        required: true
        type: string
      responses:
        "200":
          description: アンロック状態取得成功
          schema:
            $ref: '#/definitions/response.GetUnlockStatusResponse'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: このチャットにアクセスする権限がない
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: チャットが見つからない
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: アンロック状態取得
      tags:
      - matches
...
  /debug/echo:
    post:
      parameters:
      - description: メッセージ
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/requests.DebugEchoRequest'
      responses:
        "200":
          description: OK! Echo is healthy
          schema:
            $ref: '#/definitions/response.DebugEchoResponse'
      summary: Echo
      tags:
      - debug
...
  /debug/id-token:
    get:
      description: |-
        Firebase ID token（IDトークン）を直接生成します。このトークンはテスト・開発用途で使用できます。
        このトークンはBearerトークンとしてAuthorizationヘッダーで直接使用できます。
        内部的には、カスタムトークンを生成してからFirebase REST APIで交換しています。
      parameters:
      - description: Firebase User ID (UID)
        in: query
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: ID token generated successfully
          schema:
            $ref: '#/definitions/response.DebugIDTokenResponse'
        "400":
          description: userId parameter is missing or empty
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to generate ID token
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "503":
          description: Firebase Web API Key not configured
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Firebase ID Tokenを生成
      tags:
      - debug
  /diagnosis/{diagnosisId}:
    get:
      description: 特定の診断結果の詳細情報を取得する
      parameters:
      - description: 診断ID（ULID）
        in: path
        name: diagnosisId
        required: true
        type: string
      responses:
        "200":
          description: 診断詳細取得成功
          schema:
            $ref: '#/definitions/response.DiagnosisDetailResponse'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: アクセス権限がない
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 診断結果が見つからない
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: サーバーエラー
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: 診断詳細取得
      tags:
      - diagnosis
  /diagnosis/execute:
    post:
      description: 自分のAvatarと相手のAvatarの会話を分析して相性診断を実行し、マッチングポイントを加算する
      parameters:
      - description: 診断リクエスト
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.ExecuteDiagnosisRequest'
      responses:
        "200":
          description: 診断実行成功
          schema:
            $ref: '#/definitions/response.DiagnosisResult'
        "400":
          description: リクエストが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Avatarが見つからない
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: サーバーエラー
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Avatar間診断実行
      tags:
      - diagnosis
  /diagnosis/history:
    get:
      description: ユーザーの診断履歴一覧を取得する
      parameters:
      - default: 20
        description: 取得件数
        in: query
        name: limit
        type: integer
      - default: 0
        description: オフセット
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: 診断履歴取得成功
          schema:
            $ref: '#/definitions/response.DiagnosisHistoryResponse'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: サーバーエラー
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: 診断履歴取得
      tags:
      - diagnosis
...
securityDefinitions:
  Bearer:
    description: 'Firebase IDトークンをBearerトークンとして送信してください。例: "Bearer eyJhbGciOiJSUzI1NiIs..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"